requirements-weaviate.txt
weaviate-client
nltk
tqdm
requests
sentence-transformers